refactor : revised alias root and applied
