+++ src/app/api/articles/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import connectMongoDB from '@/src/libs/mongodb'
-import Article from '@/src/models/article'
+import connectMongoDB from '@/libs/mongodb'
+import Article from '@/models/article'
 
 export const POST = async (request: NextRequest) => {
   const { title, content } = await request.json()
