+++ src/app/layout.tsx
@@ -2,7 +2,8 @@ import './globals.css'
 import { Inter } from 'next/font/google'
 
 import type { Metadata } from 'next'
-import Navbar from '@/src/components/Navbar'
+
+import Navbar from '@/components/Navbar'
 
 const inter = Inter({ subsets: ['latin'] })
